requirements.txt
scikit_learn==1.5.1
beautifulsoup4
Flask
googlesearch_python
numpy
pandas
python_dateutil
requests
whois
gunicorn
bcrypt
flask-dance
flask-sqlalchemy
google-generativeai





